New script DataAnalysis/state_detector.py
state_detector.py takes a radios_with_file_output.py
script output file and determines whether the load is turned on or off
based stricly on the raw adc values for the loads current. This script
is tuned to work with our test refrigerator and Tweet-a-Watt #1, future
modification should account for more tweet-a-watts, take commmand line
arguments, and tune for other types of electric loads.